% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.R
\name{comp}
\alias{comp}
\title{Compare a recalculated to a reported value}
\usage{
comp(recalculated, reported, margin = 0.01)
}
\arguments{
\item{recalculated}{A numeric value with the recalculated result.}

\item{reported}{A character with the reported result. (Note that this is a
character and not a number. This is important to define the error margin
created by rounding. E.g., a reported value of 2.70 has a different error
margin than a reported value of 2.7)}

\item{margin}{Additional margin of error added to the margin bounds. Defaults
to 1\%.}
}
\value{
A list with different elements: \code{margin} return the error margin as
a vector. \code{recalculated} repeats the recalculated value. \code{correct} is a
logical, whether the recalculated value is within the margin.
}
\description{
Assesses the numerical correctness of a reproduced result by calculating
whether it falls into a defined margin of error around the reported number
}
\details{
The procedure for calculating the margin of error is the following: (1) Add
the the maximum possible margin of rounding, given the number of decimals,
to the value; (2) then add an additional error margin (e.g., 1\% around) this
bounds.
}
\examples{
comp(0.51, "0.5")
}
