% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcomp.R
\name{pcomp}
\alias{pcomp}
\title{Compare a recalculated p-value to a reported value p-value}
\usage{
pcomp(recalculated, reported, margin = 0.01, siglevel = 0.05)
}
\arguments{
\item{recalculated}{A numeric value with the recalculated result.}

\item{reported}{A character with the reported result. (Note that this is a
character and not a number. This is important to define the error margin
created by rounding. E.g., a reported value of 2.70 has a different error
margin than a reported value of 2.7).}

\item{margin}{Additional margin of error added to the margin bounds. Defaults
to 1\%.}

\item{siglevel}{Level of significance, to check for decision errors.}
}
\value{
A list with different elements: \code{margin} return the error margin as
a vector. \code{recalculated} repeats the recalculated value. \code{correct} is a
logical, whether the recalculated value is within the margin. \code{presult} is
a character, either 'exact', 'compatible', or 'incorrect'. \code{dec_error} is a
logical, whether a decision error is present (reported and recalculated
p-values fall on different sides of the level of significance).
}
\description{
Assesses the numerical correctness of a reproduced p-value by calculating
whether it it is exactly reported, falls within the margin of error of the
reported value, and falls on the same side of the level of significance.
}
\details{
A p-value reproduction is labeled as 'exact', when the reported p-value was
given as an exact value, the recalculated value falls within its margin of
error, and is on the same side of the level of significance (i.e. no decision
error). A p-value is labeled as 'compatible' if it was reported as an
inequality that (considering the error margin for its bound) matches the
reproduced value, and both fall on the same side of the level of
significance. A p-value reproduction is labeled as 'incorrect' otherwise.

For details on the calculation of the error margin, see \code{\link[=comp]{comp()}}.
}
\examples{
pcomp(0.51, "0.5")
pcomp(0.02, "<0.05")
pcomp(0.055, "<0.05")
}
